move indexed db inside `platform-browser`
