+++ packages/platform-browser/src/IndexedDbVersion.ts
@@ -11,7 +11,7 @@ import type * as IndexedDbTable from "./IndexedDbTable.js"
  * @category type ids
  */
 export const TypeId: unique symbol = Symbol.for(
-  "@effect/platform/IndexedDbVersion"
+  "@effect/platform-browser/IndexedDbVersion"
 )
 
 /**
