+++ packages/platform-browser/test/IndexedDbMigration.test.ts
@@ -1,4 +1,4 @@
-import { IndexedDbMigration, IndexedDbTable, IndexedDbVersion } from "@effect/platform"
+import { IndexedDbMigration, IndexedDbTable, IndexedDbVersion } from "@effect/platform-browser"
 import { assert, describe, it } from "@effect/vitest"
 import { Effect, Schema } from "effect"
