+++ packages/platform/src/index.ts
@@ -202,36 +202,6 @@ export * as HttpTraceContext from "./HttpTraceContext.js"
 /**
  * @since 1.0.0
  */
-export * as IndexedDb from "./IndexedDb.js"
-
-/**
- * @since 1.0.0
- */
-export * as IndexedDbDatabase from "./IndexedDbDatabase.js"
-
-/**
- * @since 1.0.0
- */
-export * as IndexedDbMigration from "./IndexedDbMigration.js"
-
-/**
- * @since 1.0.0
- */
-export * as IndexedDbQuery from "./IndexedDbQuery.js"
-
-/**
- * @since 1.0.0
- */
-export * as IndexedDbTable from "./IndexedDbTable.js"
-
-/**
- * @since 1.0.0
- */
-export * as IndexedDbVersion from "./IndexedDbVersion.js"
-
-/**
- * @since 1.0.0
- */
 export * as KeyValueStore from "./KeyValueStore.js"
 
 /**
